% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/file_checksum.R
\name{file_checksum}
\alias{file_checksum}
\alias{file_checksum.default}
\alias{file_checksum.webhdfs}
\title{file_checksum: Get File Checksum on a FileSystem}
\usage{
file_checksum(fs, path, ...)

\method{file_checksum}{default}(fs, path, ...)

\method{file_checksum}{webhdfs}(fs, path, ...)
}
\arguments{
\item{fs}{FileSystem object}

\item{path}{a character vector that contains the path of file}

\item{...}{other arguments}
}
\value{
a list of Checksum information
}
\description{
Get File Checksum on a FileSystem
}

