\name{read_file}
\alias{read_file}
\alias{read_file.default}
\title{read_file: Read a File on a FileSystem}
\usage{
read_file(fs, path, ...)

\method{read_file}{default}(fs, path, ...)
}
\arguments{
  \item{fs}{FileSystem object}

  \item{path}{a character vector that contains the path of
  file to read}

  \item{...}{other arguments}
}
\value{
a character vector that contains the contents of the File
}
\description{
Read a File on a FileSystem
}

