% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/write_file.R
\name{write_file}
\alias{write_file}
\alias{write_file.default}
\title{write_file: Write to a File on a FileSystem}
\usage{
write_file(fs, targetPath, ...)

\method{write_file}{default}(fs, targetPath, ...)
}
\arguments{
\item{fs}{FileSystem object}

\item{targetPath}{a character vector that contains the path of file to write}

\item{...}{other arguments}
}
\value{
\code{TRUE} if successful, \code{FALSE} otherwise
}
\description{
Write to a File on a FileSystem
}

